All menus completed & Account structure overhaul
Finished the play menu logic and all sign in logic (in the AccountManager class) All that's left here is sending the game to the levels themselves. | Finished the how to play menu. | Finished the leaderboard. | Completely overhauled the way account management works. Instead of everything in a hashmap with a nested hashmap, Account object classes are used and store din a list of Account classes.